fix(announce): poll for "agent ready" once a second
We previously polled once every ten seconds, starting immediately after
announcing ourselves. Since it takes a bit of time from the announce
until the agent has brought up the sensor for our process, there always
was a 10 second delay before we got the agent ready signal.

Now we poll once a second. At the same time, instead of 10 retries
(allowing for a maximum time of 100 seconds until we give up and restart
the full announce cycle), we now allow up to 60 retries,which gives us
a maximum time of 60 seconds before restarting the cycle. This still is
plenty.