refactor(tests): add API to express expectations as an array
Previously expectExactlyOneItem and friends (which are heavily used in
the tracing tests) only excepted a single function which contained all
relevant expectations. If it failed, it was unclear, which of the
expectations had failed, making it

(a) hard to troubleshoot failing tests locally, and
(b) making it impossible to know why exactly a test had failed on CI.

This introduces a new API that accepts an array of individual
expectations (each as its own function). Thus, we can report accurately
which item would have matched best and which expectation exactly made
the match fail.

This also migrates two arbitrary tests to the new API. Other tests will
be migrated later on.