refactor(serverless): reduce overhead for Lambda tracing
The execution time we add to do native tracing and in-process metric
collection for AWS Lambda functions is already low but we can optimize
this further by not waiting for a response from the back end.

Instead, we only wait until the request has been sent. This shaves
around 30% off the execution time that native tracing adds to a Lambda.