feat(aws-lambda): added support for SSM parameter store (#464)
refs 6483

Added support for AWS SSM.

Instead of defining the plain INSTANA_AGENT_KEY as env variable, the customer can use the parameter store.
They can set INSTANA_SSM_PARAM_NAME=/my/ssm/path. We try to fetch the value from the AWS API. The challenge here is that the API call is async.

Support for "safestring" type was added too. Customers can configure the behaviour via an env variable called INSTANA_SSM_DECRYPTION

Attention: If you set both env variables, INSTANA_AGENT_KEY is always stronger. 

The challenge of this change was that the fetch of the key is async. We had to ensure that we fetch the key as earliest as possible.

Co-authored-by: kirrg001 <katharina.irrgang@gmail.com>