test(mssql): retry connection on CI
Apparently with bad luck on CI sometimes the tests for MSSQL will start
before the MSSQL container has fully started up. This has become more
likely when we started to split tests across executors because it
increases the chance that the MSSQL tests start rather early on one of
the executors.

We now retry to connect to the MSSQL server three times when on CI,
which should solve this issue.