ci(mssql): switch to mssql express editions
In one of the recent commits, we switched to a more recent image
version of mcr.microsoft.com/mssql/server, in the hope of fixing
randomly failing MSSQL tracing integration tests. It did not help.
The container randomly crashes with exit code 137, indicating that it is
out of memory. Apparently MSSQL on Linus is quite memory hungry since
its running on a platform abstraction layer.

Unfortunately, there seems to be no way of providing more memory to just
that one container on CircleCI :-/

The documentation
(https://learn.microsoft.com/en-us/sql/linux/sql-server-linux-configure-environment-variables?view=sql-server-ver16)
indicates that you can limit the memory usage with
MSSQL_MEMORY_LIMIT_MB, but this does not have any effect with smaller
values below 2048 MB (the docs for the image say 2 GB (!) are required).

Testing locally seems to indicate that using the Express edition over
the Developer edition (which is the default) consumes slightly less
memory, so let's try that.