chore(bin): fix eslint false positive for bin directory
Importing the promises API of the core fs module via fs/promises was
only introduced in Node.js 14, which is why the eslint rule
instana/no-unsafe-require (which internally uses
require('module').builtinModules) falsely detects fs/promises as an
invalid import, when running eslint in Node.js 12.

We don't need this rule at all for bin, none of the scripts are part of
our publicly released packages. Thus we can simply disable this
particular rule for that folder.

This only came up now since we added linting for the bin folder in a
recent commit.

[skip ci]