test(prisma): do not disconnect manually from DB at end of test
We do not need this explicit disconnect step, since we are about to
terminate the Prisma integration test app at that point anyway. None of
the other DB integration test apps has such a manual disconnect step.
Furthermore, a recent CI run failed because the `after` hook of the
Prisma test suite failed, leaving the Prisma app running after the test
was finished. This blocked the test apps for all subsequent tests from
starting (because they all try to open the same port, 3215, and that
port was still blocked by the Prisma app.)