ci: use resource_class medium+ instead of large or medium
With the default resource class (medium), the MSSQL container does not
start correctly or terminates abruptly quickly no a lot of runs. We
just verified that this still happens. This is exactly the issue we have
seen earlier this year which prompted us to use "resource_class: large"
across the board. Thus, "resource_class: medium" (that is, the default
Docker executor resource class) is currently not a viable option.

Judging by a few tries, the resource_class "medium+" (which is between
medium and large) is good enough, but cheaper in terms of credit
consumption. (Tested with 2 parallel executors as well as 20, in case
that has an impact.)