ci: skip flaky Instana + otel test (#815)
This test is flaky on CI. It basically serves to demonstrate that with
this unsupported setup (combining OTel and Instana, initializing OTel
first), we might receive duplicated spans, one from OTel, one from
Instana. However, sometimes the OTel span is missing. That might be due
to a timing issue. Anyways, since this test mostly exist to document an
unsupported setup, we skip it on CI.

See also:
https://app.circleci.com/insights/github/instana/nodejs/workflows/build/tests?branch=main,
section "Flaky tests".